Improve instructions for kitty
The existing instructions for kitty are unnecessary complicated.

They also have a typo that can lead to users losing their existing configuration, by overwriting `kitty.conf` instead of writing to `colors.conf`.

We can instead take advantage of the fact that kitty supports an `include` statement in its configuration to point to the theme files directly.